Improvements to MARS Selector knob and MODIFY/CHECK RESET chord.
. Change single-stepping for the MARS Selector knob so that clicking in
the "cup" on either side of the knob moves the tip of the knob pointer
one step in the direction of the click.
. Implement a different analog for simultaneously pressing the MODIFY
and CHECK RESET buttons on the Control Panel to set up a memory clear
operation. You can now click and hold the MODIFY button, then drag to
the CHECK RESET button, and release. The original method of first
clicking MODIFY and then clicking CHECK RESET continues to work.
. Both of these were suggestions of Dave Babcock.