Correct PanelButton highlighting after mouse move.
. A PaneButton object highlights on mousedown and removes the highlight
on mouseup, animating somewhat a click. If you did mousedown and then
moved the cursor out of the button and did mouseup, a click event is
not generated, because the mouseup occurred on a different element. The
highlight remained, however, because the button never saw the mouseup.
This did not affect subsequent use of the button, however -- it was
just a cosmetic defect. The button now keeps track of the fact that it
has seen a mousedown, and if the cursor moves outside the button,
removes the highlight.
. As part of this, the click-drag-release for the MODIFY and CHECK RESET
buttons that simulates simultaneously pressing them has been improved
so that if you move outside the MODIFY button after its mousedown
(which now removes the highlight), but then drag into the CHECK RESET
button (either directly or indirectly), the down highlight for MODIFY
is reinstated.