v0.04b: Miscellaneous corrections and tweaks.
. (EXPERIMENTAL) Implement automatic carriage-return at the beginning
of and end of Typewriter INSERT operations to place the operator's
input on a separate line.
. Correct restart of the emulation clock in I/O release -- it needs to
occur after I-cycle entry, not before. This was causing the AUTOMATIC
lamp to appear to be not on.
. Implement release() method for I/O devices to handle any post-I/O
wrapup that's needed (like auto-CR after an INSERT).
. Anchor ControlPanel lamp/button section to the window left margin
instead of the right to keep it positioned properly when the window
narrows.
. Shut down the emulator when the ControlPanel is closed.