v0.08c: CardPunch and pop-up window sizing corrections.
. Correct serious CardPunch problem where the first write to the punch
when it was not ready caused the card to be put out with 81 columns and
the next card to be missing (ouch). Caused by initiateCardPunch()
exiting after waiting for the punch to become ready rather than
continuing to process the card in the buffer.
. Allow CardPunch transportReady and bufferReady signals to be canceled
when the I/O is released.
. Correct CardPunch handling of punch check and NPRO; don't clear the
internal punch buffer.
. Implement CardPunch stacker capacity as a SystemConfig parameter.
. Make CardReader/Punch annunciators slightly darker when not lit.
. Correct resizing of pop-up emulator windows to make the window size
match the size of its content.
. Change FlipFlop.lampPersistence factor to 10ms (experimental).