v0.16: Finalize floating-point implementation, plus some fixes.
. Implement new UI for Debug Options to combine toggling the Register
Debug View and the CMEM load/dump controls. This is triggered by double-
clicking the IBM log on the Control Panel and allows the Register Debug
View and AuxCEPanel to be opened and closed independently.
. Correct apparent time race in LinePrinter between clearing the
internal print line buffer and signaling that the printer was ready
for the next line of data to be transmitted from the Processor. Caused
corruption of print lines, but only noticed when running with a non-
default carriage tape.
. Correct reference to binary mask used to isolate digits for parity
checking in Processor.fetch(). This was a long-standing bug, but parity
errors never occurred, so it was never noticed.