v0.16c: Further Plotter enhancements.
. Improve printing of the plot area: now prints only the portion of the
plot area that was actually "inked". The red cursor lines are no longer
printed.
. Implement SAVE button to convert the plot area to a PNG image file.
. Implement colored pens and manual pen selection.
. Implement a system configuration option to scale the plotting area at
50% as before or 100% (with a wider window).
. Implement vertical resizing of the Plotter window and plotting area.
Horizontal resizing of the window can also be done, but that does not
resize the plotting area.
. Implement a different scrolling mechanism for the plot area. The
horizontal cursor line will now move vertically within a range of the
window, and the plot area will scroll once the cursor moves out of that
range. This is an experimental feature and may not be retained.
. Adjust the translation calculations to achieve better anti-aliasing.
. Highlight the pen-down status in red on the Plotter manual controls.
. Reduce Envir.minSliceTime for Processor throttling to improve Plotter
performance.
. Commit Plot-Demo-2.card "string art" demo.
. Correct rouge Bendix G-15 references in index.html.