Checkpoint enhancements and corrections to scrolling and resizing.
. Implement sliding of the canvas vertically to allow the cursor to stay
in the vertical middle of the plot view at low Y offsets without a
portion of the canvas equal to half the view height. Boy, was that fun
to make work properly.
. Correct problems with window resizing and the necessary plot offset
adjustments, especially as that intersected with vertical canvas
sliding.
. Correct problems with initial window sizing and compensating for some
browsers not opening windows at the size that was requested.
. Clean up instance variable nomenclature so it's more descriptive and
less confusing; clarify the difference between canvas pixels and css
(display) pixels.
. Implement brown as an additional drawing color (requested by Dave
Babcock).